update UserServices and AuthServices

+++ src/main/java/com/TrackMyItem/service/Impl/UserServiceImpl.java
@@ -2,10 +2,13 @@ package com.TrackMyItem.service.Impl;
 
 import com.TrackMyItem.dao.ItemDao;
 import com.TrackMyItem.dao.UserDao;
+import com.TrackMyItem.dao.secure.AllUsersDao;
 import com.TrackMyItem.dto.UserAllDto;
 import com.TrackMyItem.dto.UserDto;
+import com.TrackMyItem.dto.secure.AllUsersDto;
 import com.TrackMyItem.entity.ItemEntity;
 import com.TrackMyItem.entity.UserEntity;
+import com.TrackMyItem.entity.secure.AllUsersEntity;
 import com.TrackMyItem.exception.ItemNotFoundException;
 import com.TrackMyItem.exception.UserNotFoundException;
 import com.TrackMyItem.service.UserService;
@@ -25,43 +28,44 @@ public class UserServiceImpl implements UserService {
     private final UserDao userDao;
     private final EntityDtoConverter entityDtoConverter;
     private final UtilData utilData;
+    private final AllUsersDao allUsersDao;
 
     @Override
-    public void addUser(UserDto userDto) {
-        userDto.setUserId(utilData.generateUserId());
-        userDto.setLastLogin(utilData.generateTodayDate());
-        userDto.setUpdatedAt(utilData.generateTodayDate());
-        userDto.setJoinDate(utilData.generateTodayDate());
-        userDto.setRole(userDto.getRole());
-        userDao.save(entityDtoConverter.convertUserDtoToUserEntity(userDto));
+    public void updateImage(UserDto userDto) {
+        Optional<UserEntity> foundUser = userDao.findByEmail(userDto.getEmail());
+        foundUser.get().setImage(userDto.getImage());
+        foundUser.get().setUpdatedAt(utilData.generateTodayDate());
     }
 
     @Override
-    public void updateUser(String userId, UserDto userDto) {
-        Optional<UserEntity> foundUser = userDao.findById(userId);
+    public void updateUser(UserDto userDto) {
+        Optional<UserEntity> foundUser = userDao.findByEmail(userDto.getEmail());
         if(!foundUser.isPresent()) {
             throw new UserNotFoundException("User Not Found");
         }
         foundUser.get().setFirstName(userDto.getFirstName());
         foundUser.get().setLastName(userDto.getLastName());
-        foundUser.get().setEmail(userDto.getEmail());
         foundUser.get().setPhoneNumber(userDto.getPhoneNumber());
 
         foundUser.get().setUpdatedAt(utilData.generateTodayDate());
     }
-
     @Override
-    public void deleteUser(String userId) {
-        Optional<UserEntity> foundUser = userDao.findById(userId);
+    public void deleteUser(String email) {
+        Optional<UserEntity> foundUser = userDao.findByEmail(email);
         if(!foundUser.isPresent()) {
             throw new UserNotFoundException("User Not Found");
         }
+        Optional<AllUsersEntity> foundedUser = allUsersDao.findByEmail(email);
+        if(!foundedUser.isPresent()) {
+            throw new UserNotFoundException("User Not Found");
+        }
+        allUsersDao.delete(foundedUser.get());
         userDao.delete(foundUser.get());
     }
 
     @Override
-    public UserDto getUserById(String userId) {
-        Optional<UserEntity> foundUser = userDao.findById(userId);
+    public UserDto getUserByEmail(String email) {
+        Optional<UserEntity> foundUser = userDao.findByEmail(email);
         if(!foundUser.isPresent()) {
             throw new UserNotFoundException("User Not Found");
         }
